Fixed some issues in codes
