+++ src/main/java/id/ac/ui/cs/advprog/eshop/controller/ProductController.java
@@ -66,13 +66,62 @@ public class ProductController
         return "redirect:/list";
     }
 
-    @GetMapping("../")
+    @GetMapping("")
     public String homePage(Model model)
     {
         return "HomePage";
     }
 
+}
 
+@Controller
+@RequestMapping("/car")
+class CarController extends ProductController {
+    @Autowired
+    private CarServiceImpl carservice;
+
+    @GetMapping("/createCar")
+    public String createCarPage(Model model)
+    {
+        Car car = new Car();
+        model.addAttribute("car", car);
+        return "CreateCar";
+    }
+
+    @PostMapping("/createCar")
+    public String createCarPost(@ModelAttribute Car car, Model model) {
+        carservice.create(car);
+        return "redirect:listCar";
+    }
+
+    @GetMapping("/listCar")
+    public String carListPage(Model model)
+    {
+        List<Car> allCars = carservice.findAll();
+        model.addAttribute("cars", allCars);
+        return "CarList";
+    }
+
+    @GetMapping("/editCar/{carId}")
+    public String editCarPage(@PathVariable String carId, Model model)
+    {
+        Car car = carservice.findById(carId);
+        model.addAttribute("car", car);
+        return "EditCar";
+    }
+
+    @PostMapping("/editCar")
+    public String editCarPost(@ModelAttribute Car car, Model model) {
+        System.out.println(car.getCarId());
+        carservice.update(car.getCarId(), car);
+        return "redirect:listCar";
+    }
+
+    @PostMapping("/deleteCar")
+    public String deleteCar(@RequestParam("carId") String carId) {
+        carservice.deleteCarById(carId);
+        return "redirect:listCar";
+    }
 }
