+++ src/main/java/id/ac/ui/cs/advprog/eshop/repository/CarRepository.java
@@ -6,6 +6,7 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.UUID;
 
+@Repository
 public class CarRepository {
     static int id = 0;
     private List<Car> carData = new ArrayList<>();
